refactor : TabBar 패딩 수정 및 공통 Header 컴포넌트 생성
